pyproject.toml
[project]
name = "daycare-app"
version = "0.1.0"
description = ""
authors = [
    {name = "yamamoto1610",email = "1610four.season@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "requests (>=2.32.5,<3.0.0)",
    "flask-sqlalchemy (==3.1.1)",
    "flask-migrate (==4.0.7)",
    "psycopg[binary] (==3.2.3)",
    "pytz (>=2025.2,<2026.0)",
    "flask-login (==0.6.3)",
    "pillow (>=12.0.0,<13.0.0)",
    "pillow-heif (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
